requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
httpx==0.23.0
python-dotenv==0.19.0
Flask==2.0.1
flask-talisman==0.8.1
pyopenssl==23.2.0
pyjwt==2.6.0
aiohttp==3.8.1



